requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
magic-filter==1.0.12
multidict==6.0.5
packaging==24.1
pillow==10.4.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
yarl==1.9.4
